First version of music interpretation!
Right now, it works in terms of sixteenth notes. You can edit that by changing the STEP variable. Details in comments
There's a bunch of print statements for diagnostics and debugging and stuff. They're commented out so they don't run because print commands tend to slow things down a lot
I need to fix the rootdir variable. It's meant to point to the data in the project, but right now it's just a direct path to where the music is on my computer. It won't work unless you change it to match your machine.
There's a break statement in the iteration to make it just read one song right now. Trying to parse 93 songs of music with my current method produces a comically large list and takes a bit, so it's kind of overkill for testing.